+++ addOns/graphql/src/main/java/org/zaproxy/addon/graphql/GraphQlSpider.java
@@ -20,14 +20,15 @@
 package org.zaproxy.addon.graphql;
 
 import net.htmlparser.jericho.Source;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.network.HttpSender;
 import org.zaproxy.zap.spider.parser.SpiderParser;
 
 public class GraphQlSpider extends SpiderParser {
 
-    private static final Logger LOG = Logger.getLogger(GraphQlSpider.class);
+    private static final Logger LOG = LogManager.getLogger(GraphQlSpider.class);
 
     @Override
     public boolean parseResource(HttpMessage message, Source source, int depth) {
@@ -51,10 +52,10 @@ public class GraphQlSpider extends SpiderParser {
         String uri = message.getRequestHeader().getURI().toString();
         switch (MessageValidator.validate(message)) {
             case VALID_ENDPOINT:
-                LOG.debug("Found GraphQl endpoint at: " + uri);
+                LOG.debug("Found GraphQl endpoint at: {}", uri);
                 return true;
             case VALID_SCHEMA:
-                LOG.debug("Found GraphQL schema at: " + uri);
+                LOG.debug("Found GraphQL schema at: {}", uri);
                 break;
         }
         return false;