+++ addOns/graphql/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.3.0] - 2021-03-30
 ### Changed
 - Update minimum ZAP version to 2.10.0.
 - Add two new options that allow enforcing maximum query depth leniently for fields with no leaf types.
@@ -31,5 +31,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   - Import a GraphQL Schema
   - Generate Queries from an imported Schema
 
+[0.3.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.3.0
 [0.2.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.2.0
 [0.1.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.1.0